+++ packages/terra-clinical-site/src/examples/no-data-view/Index.jsx
@@ -1,31 +1,49 @@
-/* eslint-disable import/no-extraneous-dependencies */
+/* eslint-disable import/no-extraneous-dependencies, import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions */
 import React from 'react';
-import PropsTable from 'terra-props-table';
-import Markdown from 'terra-markdown';
+import DocTemplate from 'terra-doc-template';
 import ReadMe from 'terra-clinical-no-data-view/docs/README.md';
-import { version } from 'terra-clinical-no-data-view/package.json';
+import { name } from 'terra-clinical-no-data-view/package.json';
 
 // Component Source
-// eslint-disable-next-line import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions
 import NoDataViewSrc from '!raw-loader!terra-clinical-no-data-view/src/NoDataView.jsx';
 
 // Example Files
 import NoDataViewStandard from './NoDataViewStandard';
+import NoDataViewStandardSrc from '!raw-loader!./NoDataViewStandard';
 import NoDataViewContent from './NoDataViewContent';
+import NoDataViewContentSrc from '!raw-loader!./NoDataViewContent';
 import NoDataViewHiddenGlyph from './NoDataViewHiddenGlyph';
+import NoDataViewHiddenGlyphSrc from '!raw-loader!./NoDataViewHiddenGlyph';
 
-const NoDataViewExamples = () => (
-  <div>
-    <div id="version">Version: {version}</div>
-    <Markdown id="readme" src={ReadMe} />
-    <PropsTable id="props-noDataView" src={NoDataViewSrc} componentName="No Data View" />
-    <h2 id="noDataView">No Data View</h2>
-    <NoDataViewStandard />
-    <h2 id="noDataView-hidden-glyph">No Data View With Glyph Hidden</h2>
-    <NoDataViewHiddenGlyph />
-    <h2 id="noDataView-content">No Data View With Content</h2>
-    <NoDataViewContent />
-  </div>
+const DocPage = () => (
+  <DocTemplate
+    packageName={name}
+    readme={ReadMe}
+    srcPath={`https://github.com/cerner/terra-clinical/tree/master/packages/${name}`}
+    examples={[
+      {
+        title: '>No Data View',
+        example: <NoDataViewStandard />,
+        source: NoDataViewStandardSrc,
+      },
+      {
+        title: 'No Data View With Glyph Hidden',
+        example: <NoDataViewHiddenGlyph />,
+        source: NoDataViewHiddenGlyphSrc,
+      },
+      {
+        title: 'No Data View With Content',
+        example: <NoDataViewContent />,
+        source: NoDataViewContentSrc,
+      },
+    ]}
+    propsTables={[
+      {
+        componentName: 'No Data View',
+        componentSrc: NoDataViewSrc,
+      },
+    ]}
+  />
 );
 
-export default NoDataViewExamples;
+export default DocPage;