+++ packages/terra-clinical-site/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added new example for DetailView with Secondary Titles and isSmallerTitles
 
+### Changed
+* Updated site pages to use terra-doc-template
+
 2.1.0 - (March 14, 2018)
 ----------
 ### Added